+++ internal/ethapi/api.go
@@ -94,90 +94,6 @@ func (s *EthereumAPI) BaseFee(ctx context.Context) (*hexutil.Big, error) {
 	return (*hexutil.Big)(baseFee), err
 }
 
-type Price struct {
-	GasTip *hexutil.Big `json:"maxPriorityFeePerGas"`
-	GasFee *hexutil.Big `json:"maxFeePerGas"`
-}
-
-type PriceOptions struct {
-	Slow   *Price `json:"slow"`
-	Normal *Price `json:"normal"`
-	Fast   *Price `json:"fast"`
-}
-
-var (
-	minBaseFee = new(big.Int).SetUint64(1_000_000_000)   // 1 nAVAX
-	maxBaseFee = new(big.Int).SetUint64(100_000_000_000) // 100 nAVAX
-
-	minGasTip = common.Big1
-	maxGasTip = new(big.Int).SetUint64(20_000_000_000) // 100 nAVAX
-
-	slowFeeNum = new(big.Int).SetUint64(19)
-	fastFeeNum = new(big.Int).SetUint64(21)
-	feeDen     = new(big.Int).SetUint64(20)
-)
-
-// SuggestPriceOptions returns suggestions for what to display to a user for
-// current transaction fees.
-func (s *EthereumAPI) SuggestPriceOptions(ctx context.Context) (*PriceOptions, error) {
-	baseFee, err := s.b.EstimateBaseFee(ctx)
-	if err != nil || baseFee == nil {
-		return nil, err
-	}
-	gasTip, err := s.b.SuggestGasTipCap(ctx)
-	if err != nil || gasTip == nil {
-		return nil, err
-	}
-
-	slowBaseFee, normalBaseFee, fastBaseFee := priceOptions(
-		minBaseFee,
-		baseFee,
-		maxBaseFee,
-	)
-	slowGasTip, normalGasTip, fastGasTip := priceOptions(
-		minGasTip,
-		gasTip,
-		maxGasTip,
-	)
-	slowGasFee := new(big.Int).Add(slowBaseFee, slowGasTip)
-	normalGasFee := new(big.Int).Add(normalBaseFee, normalGasTip)
-	fastGasFee := new(big.Int).Add(fastBaseFee, fastGasTip)
-	return &PriceOptions{
-		Slow: &Price{
-			GasTip: (*hexutil.Big)(slowGasTip),
-			GasFee: (*hexutil.Big)(slowGasFee),
-		},
-		Normal: &Price{
-			GasTip: (*hexutil.Big)(normalGasTip),
-			GasFee: (*hexutil.Big)(normalGasFee),
-		},
-		Fast: &Price{
-			GasTip: (*hexutil.Big)(fastGasTip),
-			GasFee: (*hexutil.Big)(fastGasFee),
-		},
-	}, nil
-}
-func priceOptions(
-	minFee *big.Int,
-	estimate *big.Int,
-	maxFee *big.Int,
-) (*big.Int, *big.Int, *big.Int) {
-	// Cap the fee to keep slow and normal options reasonable during fee spikes.
-	cappedFee := math.BigMin(estimate, maxFee)
-
-	slowFee := new(big.Int).Set(cappedFee)
-	slowFee.Mul(slowFee, slowFeeNum)
-	slowFee.Div(slowFee, feeDen)
-	slowFee = math.BigMax(slowFee, minFee)
-
-	normalFee := cappedFee
-
-	fastFee := new(big.Int).Set(estimate)
-	fastFee.Mul(fastFee, fastFeeNum)
-	fastFee.Div(fastFee, feeDen)
-	return slowFee, normalFee, fastFee
-}
-
 // MaxPriorityFeePerGas returns a suggestion for a gas tip cap for dynamic fee transactions.
 func (s *EthereumAPI) MaxPriorityFeePerGas(ctx context.Context) (*hexutil.Big, error) {
 	tipcap, err := s.b.SuggestGasTipCap(ctx)
